requirements.txt
# ChannelSmith - Python Dependencies

# Core Dependencies
Pillow>=10.0.0          # Image processing library
numpy>=1.24.0           # Numerical computing

# Web UI Dependencies
Flask>=3.0.0            # Web framework
flask-cors>=4.0.0       # CORS support for Flask

# Development Dependencies
pytest>=7.4.0           # Testing framework
pytest-cov>=4.1.0       # Test coverage reporting
black>=23.0.0           # Code formatter
pylint>=2.17.0          # Code linter

# Future Dependencies (commented out for now, will add in later phases)
# pyinstaller>=5.13.0   # For creating .exe (RC phase)
